Moved pkg files and updated refs
