+++ pkg/processor/attraction/attraction.go
