+++ pkg/processor/attraction/attraction_test.go
