+++ pkg/processor/attraction/hasher.go
